[CELEBORN-2096]Support Lz4 Compression in CppClient

+++ cpp/celeborn/client/compress/Compressor.h
@@ -32,7 +32,7 @@ class Compressor {
 
   virtual void initCompressBuffer(int maxDestLength) = 0;
   virtual void compress(const uint8_t* data, int offset, int length) = 0;
-  virtual int getCompressedTotalSize() const = 0;
+  virtual size_t getCompressedTotalSize() const = 0;
   virtual const std::vector<uint8_t>& getCompressedBuffer() const = 0;
 
   static std::unique_ptr<Compressor> createCompressor(
